backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
httpx>=0.27.0
sentence-transformers>=2.5.0
faiss-cpu>=1.7.4
qdrant-client>=1.9.0
pypdf>=4.0.0
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-cov>=5.0.0
